We have channels \o/

+++ app/src/main/java/com/marcsello/matterless/model/Channels.kt
@@ -14,5 +14,5 @@ data class Channels (
     @SerializedName("name")
     var name: String? = null,
     @SerializedName("last_post_at")
-    var lastPostat: Int? = null
+    var lastPostat: Long? = null
 )
